refactor: consolidate conversion tracking functionality into modular hooks for improved maintainability and clarity
